pyproject.toml
[project]
name = "pymultiastar"
version = "0.0.6"
description="Multi-goal A* with cpp bindings to python"
readme = "README.md"
authors = [
  { name = "Jeremy Castagno", email = "jeremybyu@gmail.com" },
]
requires-python = ">=3.9"
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11"
]
dependencies = [
  "numpy>=1.15",
]

[build-system]
requires = ["scikit-build-core>=0.3.3", "pybind11"]
build-backend = "scikit_build_core.build"

[project.optional-dependencies]
geo = ["pyproj>=3.0"]
test = ["pytest", "Pillow", "pyproj>=3.0"]
docs = ["mkdocstrings[python]", "mkdocs-autorefs>=0.3.1",
      "pymdown-extensions>=6.3", "mkdocs>=1.3", "griffe>=0.27.5", "python-markdown-math",
      "mkdocs-coverage>=0.2", "mkdocs-gen-files>=0.3",
      "mkdocs-literate-nav>=0.4","mkdocs-material>=7.3",
      "mkdocs-section-index>=0.3", "mkdocstrings-python>=0.5.1",
      "markdown-callouts>=0.2","markdown-exec>=0.5"]

dev = ["black", "open3d>=0.17", "matplotlib", 
      "rich", "typer", "ipdb", "Pillow", "pandas", "seaborn",
      "pymultiastar[geo,test]"]


[tool.scikit-build]
wheel.expand-macos-universal-tags = true

[tool.cibuildwheel]
test-command = "pytest {project}/tests"
test-extras = ["test"]
test-skip = ["*universal2:arm64"]
build-verbosity = 1

[tool.setuptools]
package-dir = {"" = "src"}

[tool.ruff]
line-length = 100
target-version = "py310"